product-service/requirements.txt
flask==3.0.3
mysql-connector-python==8.4.0
redis==5.0.7